Ajustando definição das colunas com o banco de dados

+++ src/main/java/br/com/fiap/spacecrop/dto/request/FazendaRequestDTO.java
@@ -27,5 +27,5 @@ public class FazendaRequestDTO {
 
     @NotNull(message = "Área em hectares é obrigatória")
     @Positive(message = "Área deve ser maior que zero")
-    private Float areaHectares;
+    private Double areaHectares;
 }
